ENH: block decomposition for beta diversity
