+++ skbio/diversity/__init__.py
@@ -163,6 +163,7 @@ Functions
     alpha_diversity
     beta_diversity
     partial_beta_diversity
+    block_beta_diversity
     get_alpha_diversity_metrics
     get_beta_diversity_metrics
 
@@ -402,8 +403,10 @@ from skbio.util import TestRunner
 
 from ._driver import (alpha_diversity, beta_diversity, partial_beta_diversity,
                       get_alpha_diversity_metrics, get_beta_diversity_metrics)
+from ._block import block_beta_diversity
 
 __all__ = ["alpha_diversity", "beta_diversity", "get_alpha_diversity_metrics",
-           "get_beta_diversity_metrics", "partial_beta_diversity"]
+           "get_beta_diversity_metrics", "partial_beta_diversity",
+           "block_beta_diversity"]
 
 test = TestRunner(__file__).test
